+++ fake-cucumber/javascript/src/Cucumber.ts
@@ -4,6 +4,8 @@ import TestPlan from './TestPlan'
 import IStepDefinition from './IStepDefinition'
 import IHook from './IHook'
 import { GherkinQuery } from 'gherkin'
+import { ParameterType } from 'cucumber-expressions'
+import IClock from './IClock'
 
 export default class Cucumber {
   constructor(
@@ -11,16 +13,21 @@ export default class Cucumber {
     // * GherkinDocument (sent through)
     // * Pickle (used)
     private readonly gherkinMessages: messages.IEnvelope[],
+    private readonly parameterTypes: Array<ParameterType<any>>,
     private readonly stepDefinitions: IStepDefinition[],
     private readonly beforeHooks: IHook[],
     private readonly afterHooks: IHook[],
     private readonly gherkinQuery: GherkinQuery,
-    private readonly newId: IdGenerator.NewId
+    private readonly newId: IdGenerator.NewId,
+    private readonly clock: IClock
   ) {}
 
   public async execute(notifier: MessageNotifier): Promise<void> {
     for (const gherkinMessage of this.gherkinMessages) {
       notifier(gherkinMessage)
+    }
+    for (const parameterType of this.parameterTypes) {
+      notifier(parameterTypeToMessage(parameterType))
     }
     for (const stepDefinition of this.stepDefinitions) {
       notifier(stepDefinition.toMessage())
@@ -37,8 +44,22 @@ export default class Cucumber {
       this.beforeHooks,
       this.afterHooks,
       this.gherkinQuery,
-      this.newId
+      this.newId,
+      this.clock
     )
     await testPlan.execute(notifier)
   }
+}
+
+function parameterTypeToMessage(
+  parameterType: ParameterType<any>
+): messages.IEnvelope {
+  return new messages.Envelope({
+    parameterType: new messages.ParameterType({
+      name: parameterType.name,
+      regularExpressions: parameterType.regexpStrings,
+      preferForRegularExpressionMatch: parameterType.preferForRegexpMatch,
+      useForSnippets: parameterType.useForSnippets,
+    }),
+  })
 }