+++ fake-cucumber/javascript/src/SupportCode.ts
@@ -1,5 +1,6 @@
 import {
   CucumberExpression,
+  ParameterType,
   ParameterTypeRegistry,
   RegularExpression,
 } from 'cucumber-expressions'
@@ -10,6 +11,7 @@ import ExpressionStepDefinition from './ExpressionStepDefinition'
 import IStepDefinition from './IStepDefinition'
 import IHook from './IHook'
 import Hook from './Hook'
+import IClock from './IClock'
 
 type RegisterStepDefinition = (
   expression: string | RegExp,
@@ -17,6 +19,19 @@ type RegisterStepDefinition = (
 ) => void
 
 type RegisterHook = (tagExpression: string, body: AnyBody) => void
+
+interface IParameterTypeDefinition {
+  name: string
+  regexp: RegExp | RegExp[] | string | string[]
+  type: any
+  transformer?: (...args: string[]) => any
+  preferForRegexpMatch?: boolean
+  useForSnippets?: boolean
+}
+
+function defaultTransformer(...args: string[]) {
+  return args
+}
 
 function getSourceReference(stackTrace: string): messages.ISourceReference {
   const stack = new StackUtils({
@@ -38,6 +53,7 @@ function getSourceReference(stackTrace: string): messages.ISourceReference {
  * This class provides an API for defining step definitions and hooks.
  */
 export default class SupportCode {
+  public readonly parameterTypes: Array<ParameterType<any>> = []
   public readonly stepDefinitions: IStepDefinition[] = []
   public readonly beforeHooks: IHook[] = []
   public readonly afterHooks: IHook[] = []
@@ -54,7 +70,22 @@ export default class SupportCode {
   public readonly After = this.registerAfterHook.bind(this) as RegisterHook
   private readonly parameterTypeRegistry = new ParameterTypeRegistry()
 
-  constructor(public newId: IdGenerator.NewId) {}
+  constructor(public newId: IdGenerator.NewId, public clock: IClock) {}
+
+  public defineParameterType(
+    parameterTypeDefinition: IParameterTypeDefinition
+  ) {
+    const parameterType = new ParameterType<any>(
+      parameterTypeDefinition.name,
+      parameterTypeDefinition.regexp,
+      parameterTypeDefinition.type,
+      parameterTypeDefinition.transformer || defaultTransformer,
+      parameterTypeDefinition.useForSnippets,
+      parameterTypeDefinition.preferForRegexpMatch
+    )
+    this.parameterTypeRegistry.defineParameterType(parameterType)
+    this.parameterTypes.push(parameterType)
+  }
 
   private registerStepDefinition(
     expression: string | RegExp,