+++ fake-cucumber/javascript/src/index.ts
@@ -1,4 +1,5 @@
 import SupportCode from './SupportCode'
 import { IdGenerator } from 'cucumber-messages'
+import PerfHooksClock from './PerfHooksClock'
 
-export = new SupportCode(IdGenerator.uuid())
+export = new SupportCode(IdGenerator.uuid(), new PerfHooksClock())