+++ fake-cucumber/javascript/src/makePickleTestStep.ts
@@ -2,12 +2,14 @@ import { messages } from 'cucumber-messages'
 import IStepDefinition from './IStepDefinition'
 import PickleTestStep from './PickleTestStep'
 import ITestStep from './ITestStep'
+import IClock from './IClock'
 
 export default function makePickleTestStep(
   testStepId: string,
   pickleStep: messages.Pickle.IPickleStep,
   stepDefinitions: IStepDefinition[],
-  sourceFrames: string[]
+  sourceFrames: string[],
+  clock: IClock
 ): ITestStep {
   const supportCodeExecutors = stepDefinitions
     .map(stepDefinition => stepDefinition.match(pickleStep))
@@ -17,6 +19,7 @@ export default function makePickleTestStep(
     pickleStep.id,
     false,
     supportCodeExecutors,
-    sourceFrames
+    sourceFrames,
+    clock
   )
 }