+++ fake-cucumber/javascript/test/makeTestCaseTest.ts
@@ -5,6 +5,7 @@ import ExpressionStepDefinition from '../src/ExpressionStepDefinition'
 import Hook from '../src/Hook'
 import { CucumberExpression, ParameterTypeRegistry } from 'cucumber-expressions'
 import { GherkinQuery } from 'gherkin'
+import IncrementClock from '../src/IncrementClock'
 
 describe('makeTestCase', () => {
   it('transforms a Pickle to a TestCase', () => {
@@ -16,7 +17,8 @@ describe('makeTestCase', () => {
       [],
       [],
       new GherkinQuery(),
-      IdGenerator.incrementing()
+      IdGenerator.incrementing(),
+      new IncrementClock()
     )
 
     assert.deepStrictEqual(
@@ -37,7 +39,8 @@ describe('makeTestCase', () => {
           beforeHooks,
           [],
           new GherkinQuery(),
-          IdGenerator.incrementing()
+          IdGenerator.incrementing(),
+          new IncrementClock()
         )
 
         assert.deepStrictEqual(
@@ -63,7 +66,8 @@ describe('makeTestCase', () => {
         [],
         afterHooks,
         new GherkinQuery(),
-        IdGenerator.incrementing()
+        IdGenerator.incrementing(),
+        new IncrementClock()
       )
 
       assert.deepStrictEqual(